# improve performance of unique with inverse indices