@soumith 2019-01-18
before I land this, could you please locally check correctness as well, against CPU.

The unit-tests only cover small_3d test cases.