@ngimel 2019-01-23
Yes, it's possible, my original commit was returning `unique_index` instead of `inverse_index`, so just restore thrust::unique_by_key call with appropriate handling of options? https://github.com/pytorch/pytorch/pull/16145/commits/883edac17355bc455f64fce7d4b8c93984eeae0a